requirements.txt
black
pylint
colorama
pygithub
python-dotenv